+++ Conf.Designer.cs
@@ -85,7 +85,7 @@ namespace TaskBarDragAndDrop {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("running window")]
+        [global::System.Configuration.DefaultSettingValueAttribute("runningwindow")]
         public string RunningWin {
             get {
                 return ((string)(this["RunningWin"]));
@@ -121,7 +121,7 @@ namespace TaskBarDragAndDrop {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("running windows")]
+        [global::System.Configuration.DefaultSettingValueAttribute("runningwindows")]
         public string multiWin {
             get {
                 return ((string)(this["multiWin"]));